pyproject.toml
[project]
name = "etail-marketplaces-sdk"
version = "0.1.0"
description = "Unified SDK for eCommerce marketplace and aggregator APIs — fetch, normalise, and sink data streams (orders, stock, catalogue, shipments, returns, invoices, analytics, ads, settlements, reviews)"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
]

[project.optional-dependencies]
postgres = ["sqlalchemy>=2.0", "psycopg2-binary>=2.9"]
supabase = ["supabase>=2.0"]
bigquery = ["google-cloud-bigquery>=3.0"]
all = [
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "supabase>=2.0",
    "google-cloud-bigquery>=3.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "bump-my-version>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["etail_marketplaces_sdk"]

[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "chore: bump version {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
file = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
file = "etail_marketplaces_sdk/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'
